requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
attrs==22.1.0
    # via pytest
billiard==3.6.4.0
    # via celery
boto3==1.24.96
    # via -r requirements.in
botocore==1.27.96
    # via
    #   boto3
    #   s3transfer
celery==5.2.7
    # via -r requirements.in
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django==4.1.2
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-filter==22.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
drf-yasg==1.21.4
    # via -r requirements.in
factory-boy==3.2.1
    # via pytest-factoryboy
faker==15.1.1
    # via
    #   -r requirements.in
    #   factory-boy
idna==3.4
    # via requests
inflection==0.5.1
    # via
    #   drf-yasg
    #   pytest-factoryboy
iniconfig==1.1.1
    # via pytest
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.2.4
    # via celery
markdown==3.4.1
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
packaging==21.3
    # via
    #   drf-yasg
    #   pytest
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.31
    # via click-repl
py==1.11.0
    # via pytest
pygments==2.13.0
    # via -r requirements.in
pyjwt==1.7.1
    # via djangorestframework-simplejwt
pyparsing==3.0.9
    # via packaging
pytest==7.1.3
    # via
    #   pytest-django
    #   pytest-factoryboy
pytest-django==4.5.2
    # via -r requirements.in
pytest-factoryboy==2.5.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   faker
python-decouple==3.6
    # via -r requirements.in
python-environ==0.4.54
    # via -r requirements.in
pytz==2022.4
    # via
    #   celery
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0
    # via -r requirements.in
requests==2.28.1
    # via coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via pytest
typing-extensions==4.4.0
    # via pytest-factoryboy
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   coreapi
    #   drf-yasg
urllib3==1.26.12
    # via
    #   botocore
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
